+++ src/Shark.Fido2.Core/Abstractions/IAttestation.cs
@@ -1,4 +1,5 @@
 ﻿using Shark.Fido2.Domain;
+using Shark.Fido2.Domain.Options;
 
 namespace Shark.Fido2.Core.Abstractions;
