+++ src/Shark.Fido2.Core/Assertion.cs
@@ -9,6 +9,7 @@ using Shark.Fido2.Core.Configurations;
 using Shark.Fido2.Domain;
 using Shark.Fido2.Domain.Constants;
 using Shark.Fido2.Domain.Enums;
+using Shark.Fido2.Domain.Options;
 
 namespace Shark.Fido2.Core;
