+++ src/Shark.Fido2.Core/Attestation.cs
@@ -8,6 +8,7 @@ using Shark.Fido2.Core.Constants;
 using Shark.Fido2.Domain;
 using Shark.Fido2.Domain.Constants;
 using Shark.Fido2.Domain.Enums;
+using Shark.Fido2.Domain.Options;
 
 namespace Shark.Fido2.Core;
