+++ src/Shark.Fido2.Core/Handlers/AssertionObjectHandler.cs
@@ -4,6 +4,7 @@ using Shark.Fido2.Core.Abstractions.Services;
 using Shark.Fido2.Core.Abstractions.Validators;
 using Shark.Fido2.Core.Results.Attestation;
 using Shark.Fido2.Domain;
+using Shark.Fido2.Domain.Options;
 
 namespace Shark.Fido2.Core.Handlers;
