+++ src/Shark.Fido2.Core/Mappers/Ec2KeyTypeMapper.cs
@@ -1,7 +1,8 @@
 ﻿using System.Security.Cryptography;
+using Shark.Fido2.Domain;
 using Shark.Fido2.Domain.Enums;
 
-namespace Shark.Fido2.Domain.Mappers;
+namespace Shark.Fido2.Core.Mappers;
 
 public static class Ec2KeyTypeMapper
 {
