+++ src/Shark.Fido2.Core/Validators/AssertionResponseValidator.cs
@@ -8,6 +8,7 @@ using Shark.Fido2.Core.Helpers;
 using Shark.Fido2.Core.Results;
 using Shark.Fido2.Domain;
 using Shark.Fido2.Domain.Enums;
+using Shark.Fido2.Domain.Options;
 
 namespace Shark.Fido2.Core.Validators;
