+++ src/Shark.Fido2.Core/Validators/AttestationStatementValidators/TpmAttestationStatementStrategy.cs
@@ -3,10 +3,10 @@ using Shark.Fido2.Core.Abstractions.Validators.AttestationStatementValidators;
 using Shark.Fido2.Core.Comparers;
 using Shark.Fido2.Core.Constants;
 using Shark.Fido2.Core.Helpers;
+using Shark.Fido2.Core.Mappers;
 using Shark.Fido2.Core.Results;
 using Shark.Fido2.Domain;
 using Shark.Fido2.Domain.Enums;
-using Shark.Fido2.Domain.Mappers;
 using Shark.Fido2.Domain.Tpm;
 
 namespace Shark.Fido2.Core.Validators.AttestationStatementValidators;
