+++ src/Shark.Fido2.Domain/Extensions/AuthenticationExtensionsClientInputs.cs
