+++ src/Shark.Fido2.Domain/Extensions/AuthenticationExtensionsClientOutputs.cs
@@ -1,4 +1,6 @@
-﻿namespace Shark.Fido2.Domain;
+﻿using Shark.Fido2.Domain.Extensions;
+
+namespace Shark.Fido2.Domain;
 
 public sealed class AuthenticationExtensionsClientOutputs
 {
