+++ src/Shark.Fido2.Domain/Extensions/CredentialPropertiesOutput.cs
@@ -1,4 +1,4 @@
-﻿namespace Shark.Fido2.Domain;
+﻿namespace Shark.Fido2.Domain.Extensions;
 
 public sealed class CredentialPropertiesOutput
 {
