+++ src/Shark.Fido2.Domain/Options/AuthenticatorSelectionCriteria.cs
@@ -1,6 +1,6 @@
 ﻿using Shark.Fido2.Domain.Enums;
 
-namespace Shark.Fido2.Domain;
+namespace Shark.Fido2.Domain.Options;
 
 /// <summary>
 /// 5.4.4. Authenticator Selection Criteria
