+++ src/Shark.Fido2.Domain/Options/PublicKeyCredentialCreationOptionsRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace Shark.Fido2.Domain;
+﻿namespace Shark.Fido2.Domain.Options;
 
 public sealed class PublicKeyCredentialCreationOptionsRequest
 {
