+++ src/Shark.Fido2.Domain/Options/PublicKeyCredentialUserEntity.cs
@@ -1,4 +1,4 @@
-﻿namespace Shark.Fido2.Domain;
+﻿namespace Shark.Fido2.Domain.Options;
 
 /// <summary>
 /// 5.4.3. User Account Parameters for Credential Generation (dictionary PublicKeyCredentialUserEntity)
