+++ src/Shark.Fido2.Models/Mappers/PublicKeyCredentialAttestationMapper.cs
@@ -1,6 +1,7 @@
 ﻿using Shark.Fido2.Common.Extensions;
 using Shark.Fido2.Domain;
 using Shark.Fido2.Domain.Enums;
+using Shark.Fido2.Domain.Extensions;
 using Shark.Fido2.Models.Responses;
 
 namespace Shark.Fido2.Models.Mappers;
