+++ src/Shark.Fido2.Models/Mappers/PublicKeyCredentialCreationOptionsRequestMapper.cs
@@ -1,6 +1,6 @@
 ﻿using Shark.Fido2.Common.Extensions;
-using Shark.Fido2.Domain;
 using Shark.Fido2.Domain.Enums;
+using Shark.Fido2.Domain.Options;
 using Shark.Fido2.Models.Requests;
 
 namespace Shark.Fido2.Models.Mappers;
