+++ src/Shark.Fido2.Models/Mappers/PublicKeyCredentialRequestOptionsMapper.cs
@@ -1,5 +1,6 @@
 ﻿using Shark.Fido2.Common.Extensions;
 using Shark.Fido2.Domain;
+using Shark.Fido2.Domain.Options;
 using Shark.Fido2.Models.Responses;
 
 namespace Shark.Fido2.Models.Mappers;
