+++ tests/Shark.Fido2.Core.Tests/AssertionTests.cs
@@ -11,6 +11,7 @@ using Shark.Fido2.Core.Results.Attestation;
 using Shark.Fido2.Domain;
 using Shark.Fido2.Domain.Constants;
 using Shark.Fido2.Domain.Enums;
+using Shark.Fido2.Domain.Options;
 
 namespace Shark.Fido2.Core.Tests;
