+++ tests/Shark.Fido2.Core.Tests/Handlers/AttestationObjectHandlerTests.cs
@@ -5,6 +5,7 @@ using Shark.Fido2.Core.Handlers;
 using Shark.Fido2.Core.Results;
 using Shark.Fido2.Core.Services;
 using Shark.Fido2.Domain;
+using Shark.Fido2.Domain.Options;
 
 namespace Shark.Fido2.Core.Tests.Handlers;
