+++ tests/Shark.Fido2.Core.Tests/PublicKeyCredentialCreationOptionsBuilder.cs
@@ -1,6 +1,7 @@
 ﻿using Shark.Fido2.Domain;
 using Shark.Fido2.Domain.Constants;
 using Shark.Fido2.Domain.Enums;
+using Shark.Fido2.Domain.Options;
 
 namespace Shark.Fido2.Core.Tests;
