+++ tests/Shark.Fido2.Core.Tests/Validators/AttestationObjectValidatorTests.cs
@@ -8,6 +8,7 @@ using Shark.Fido2.Core.Services;
 using Shark.Fido2.Core.Validators;
 using Shark.Fido2.Domain;
 using Shark.Fido2.Domain.Enums;
+using Shark.Fido2.Domain.Options;
 using Shark.Fido2.Metadata.Core.Abstractions;
 using Shark.Fido2.Metadata.Core.Domain;
