+++ tests/Shark.Fido2.Core.Tests/Validators/AttestationStatementValidators/NoneAttestationStatementStrategyTests.cs
@@ -7,6 +7,7 @@ using Shark.Fido2.Core.Tests.DataReaders;
 using Shark.Fido2.Core.Validators.AttestationStatementValidators;
 using Shark.Fido2.Domain;
 using Shark.Fido2.Domain.Enums;
+using Shark.Fido2.Domain.Options;
 
 namespace Shark.Fido2.Core.Tests.Validators.AttestationStatementValidators;
