+++ tests/Shark.Fido2.Core.Tests/Validators/AttestationStatementValidators/PackedAttestationStatementStrategyTests.cs
@@ -9,6 +9,7 @@ using Shark.Fido2.Core.Validators;
 using Shark.Fido2.Core.Validators.AttestationStatementValidators;
 using Shark.Fido2.Domain;
 using Shark.Fido2.Domain.Enums;
+using Shark.Fido2.Domain.Options;
 
 namespace Shark.Fido2.Core.Tests.Validators.AttestationStatementValidators;
